+++ cpp/src/jbu.cpp
@@ -15,8 +15,8 @@ extern "C"
     void vjbu_set_jump(Vjbu *dut, uint8_t val) { dut->jump = val; }
     void vjbu_set_branch(Vjbu *dut, uint8_t val) { dut->branch = val; }
     void vjbu_set_funct3(Vjbu *dut, uint8_t val) { dut->funct3 = val; }
-    void vjbu_set_rs1_data(Vjbu *dut, uint32_t val) { dut->rs1_data = val; }
-    void vjbu_set_rs2_data(Vjbu *dut, uint32_t val) { dut->rs2_data = val; }
+    void vjbu_set_rs1_data(Vjbu *dut, uint64_t val) { dut->rs1_data = val; }
+    void vjbu_set_rs2_data(Vjbu *dut, uint64_t val) { dut->rs2_data = val; }
 
     // Port Getters
     uint8_t vjbu_get_je(Vjbu *dut) { return dut->je; }
