+++ cpp/src/jbu.rs
@@ -7,8 +7,8 @@ unsafe extern "C" {
     pub fn vjbu_set_jump(dut: *mut std::ffi::c_void, val: u8);
     pub fn vjbu_set_branch(dut: *mut std::ffi::c_void, val: u8);
     pub fn vjbu_set_funct3(dut: *mut std::ffi::c_void, val: u8);
-    pub fn vjbu_set_rs1_data(dut: *mut std::ffi::c_void, val: u32);
-    pub fn vjbu_set_rs2_data(dut: *mut std::ffi::c_void, val: u32);
+    pub fn vjbu_set_rs1_data(dut: *mut std::ffi::c_void, val: u64);
+    pub fn vjbu_set_rs2_data(dut: *mut std::ffi::c_void, val: u64);
 
     // Getters
     pub fn vjbu_get_je(dut: *mut std::ffi::c_void) -> u8;
